+++ change_log.md
@@ -1,9 +1,14 @@
 ##2.0.4
 **New:**
 
+- Excel importing. For more information, visit the Wiki.
+- Custom error dialogs throughout the application.
+- New Row Tool column headers to display more info on mouse hover. To disable this, you may go to Settings, then switch on the new option for "Sticky Column Headers".
+- Numerous style enhancements throughout.
+
 **Bugs:**
 
-- Java Code tool should now generate proper code for Yarn-Core 0.3.4.
+- Java Code tool should now generate proper code for Yari-Core 0.3.4.
 - Correction of 33 code smells found by Sonarqube.
 
 **Issues:**